completed Exercise 4.21*: Blog List Expansion, step 9

+++ part4/controllers/blogs.js
@@ -41,8 +41,23 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'invalid user' })
+    }
+    const userid = decodedToken.id
+
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(404).json({ error: 'blog not found' })
+    }
+
+    if (blog.user._id.toString() === userid.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).end()
+    } else {
+        return response.status(403).json({ error: 'user not authorized' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
